Stop declaring/using dates
Due to timezone issues, some of my past posts might have been incorrect:
I had realized a while back that I needed to use UTC, but I'm not sure I
always did. I attempted to compare `date` strings in the YAML front
matter with Git commit times, but it's hard to determine exactly whether
each timestamp is correct. It's simplest to just stop reporting times
and stick with dates.

If I post two times in one day, I can use the `date` to disambiguate.
This is (sort of) a suggestion in the [Jekyll documentation][1].

[1]: http://jekyllrb.com/docs/frontmatter/